Update VatRateSums.cs
float to decimal

+++ SzamlazzHuSDK/Model/VatRateSums.cs
@@ -2,12 +2,12 @@
 {
     public class VatRateSums
     {
-        public float VatRate { get; set; }
+        public decimal VatRate { get; set; }
 
-        public float Net { get; set; }
+        public decimal Net { get; set; }
 
-        public float Vat { get; set; }
+        public decimal Vat { get; set; }
 
-        public float Gross { get; set; }
+        public decimal Gross { get; set; }
     }
 }
